fixup! include error_msg for srv
